# frontend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/applications/me/gap-preview": {
      "get": {
        "description": "Returns gap analysis BEFORE applying.\n\nDoes not persist data.",
        "operationId": "ApplicantsController_getGapPreview",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "query",
            "description": "Job to preview gap analysis against",
            "schema": {
              "example": "job_123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Gap preview with updated scorecard, skills gap, and review outcome.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "jobId": "job_123",
                      "fitTier": "PROBE",
                      "reviewOutcome": "NEEDS_REVIEW",
                      "roleFitScore": 72,
                      "gapSummary": "Review-recommended candidate. High backend capability. Missing Rust experience.",
                      "matchedTechnologies": [
                        "Solana"
                      ],
                      "missingTechnologies": [
                        "Rust"
                      ],
                      "skillsGap": {
                        "requiredSkills": [
                          "Rust",
                          "Solana"
                        ],
                        "matchedTechnologies": [
                          "Solana"
                        ],
                        "missingTechnologies": [
                          "Rust"
                        ],
                        "gaps": [
                          {
                            "skill": "Rust",
                            "severity": "DEALBREAKER"
                          }
                        ]
                      },
                      "scorecard": {
                        "summary": "Backend-focused developer with Solana activity.",
                        "capabilities": {
                          "backend": {
                            "score": 0.82,
                            "confidence": "high"
                          },
                          "frontend": {
                            "score": 0.42,
                            "confidence": "medium"
                          },
                          "devops": {
                            "score": 0.51,
                            "confidence": "medium"
                          }
                        },
                        "ownership": {
                          "ownedProjects": 5,
                          "activelyMaintained": 3,
                          "confidence": "high"
                        },
                        "impact": {
                          "activityLevel": "high",
                          "consistency": "strong",
                          "externalContributions": 12,
                          "confidence": "high"
                        },
                        "reputation": null,
                        "organizations": [],
                        "interactionProfile": null,
                        "stack": {
                          "languages": [
                            "TypeScript"
                          ],
                          "tools": [
                            "Anchor"
                          ]
                        },
                        "web3": {
                          "ecosystem": "solana",
                          "ecosystemPRs": 2
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing jobId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Preview gap analysis",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/me/{jobId}": {
      "post": {
        "description": "Allows an authenticated candidate to apply for an ACTIVE job.\n\nThis triggers:\n- Gap analysis\n- Decision card generation\n- Initial pipeline stage = APPLIED",
        "operationId": "ApplicantsController_apply",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "description": "Job ID",
            "schema": {
              "example": "job_123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Application submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Already applied OR missing analysis",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Apply for a job",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/me": {
      "get": {
        "description": "Returns candidate applications with human-readable pipeline stages.",
        "operationId": "ApplicantsController_getMyApplications",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Get my applications",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/hr/jobs/{jobId}": {
      "get": {
        "description": "Returns all applications for a job including:\n- HR decision view\n- Technical evaluation\n- Candidate info\n\nSupports filtering.",
        "operationId": "ApplicantsController_getJobApplications",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "job_123",
              "type": "string"
            }
          },
          {
            "name": "fitTier",
            "required": false,
            "in": "query",
            "description": "Filter by fit tier",
            "schema": {
              "enum": [
                "STRONG",
                "PROBE",
                "PASS"
              ],
              "type": "string"
            }
          },
          {
            "name": "minScore",
            "required": false,
            "in": "query",
            "description": "Minimum role fit score",
            "schema": {
              "example": 70,
              "type": "number"
            }
          },
          {
            "name": "pipelineStage",
            "required": false,
            "in": "query",
            "description": "Filter by pipeline stage",
            "schema": {
              "enum": [
                "APPLIED",
                "REVIEWED",
                "SHORTLISTED",
                "INTERVIEW_HR",
                "INTERVIEW_TECHNICAL",
                "INTERVIEW_FINAL",
                "OFFER",
                "HIRED",
                "REJECTED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Applications retrieved successfully"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "List job applications (HR)",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/hr/{appId}": {
      "get": {
        "description": "Returns full application data including:\n- Decision card\n- Gap report\n- Candidate profiles\n- Interview questions",
        "operationId": "ApplicantsController_getApplicationDetail",
        "parameters": [
          {
            "name": "appId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "app_123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns application detail with dual-view shape (HR vs CTO)",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "app_123",
                    "hrView": {
                      "verdict": "PROCEED",
                      "reviewOutcome": "OK",
                      "hrSummary": "Strong candidate with clear career growth",
                      "reputationNote": "Verified by 5 peers on-chain",
                      "appliedAt": "2026-04-20T10:00:00Z",
                      "pipelineStage": "INTERVIEW_TECHNICAL",
                      "candidate": {
                        "name": "Alice Smith",
                        "username": "alice_dev"
                      }
                    },
                    "technicalView": {
                      "technicalSummary": "Expert in React and Node.js. Missing Rust experience.",
                      "strengths": [
                        "Architecture",
                        "Testing"
                      ],
                      "risks": [
                        "No experience with high-scale DBs"
                      ],
                      "roleFitScore": 85,
                      "fitTier": "STRONG",
                      "scorecard": {
                        "summary": "Full-stack developer with strong backend capability.",
                        "capabilities": {
                          "backend": {
                            "score": 0.82,
                            "confidence": "high"
                          },
                          "frontend": {
                            "score": 0.68,
                            "confidence": "medium"
                          },
                          "devops": {
                            "score": 0.44,
                            "confidence": "medium"
                          }
                        },
                        "ownership": {
                          "ownedProjects": 5,
                          "activelyMaintained": 3,
                          "confidence": "medium"
                        },
                        "impact": {
                          "activityLevel": "high",
                          "consistency": "strong",
                          "externalContributions": 12,
                          "confidence": "high"
                        },
                        "reputation": null,
                        "organizations": [],
                        "interactionProfile": null,
                        "stack": {
                          "languages": [
                            "TypeScript"
                          ],
                          "tools": [
                            "NestJS",
                            "PostgreSQL"
                          ]
                        },
                        "web3": null
                      },
                      "skillsGap": {
                        "requiredSkills": [
                          "Rust",
                          "Solana"
                        ],
                        "matchedTechnologies": [
                          "Solana"
                        ],
                        "missingTechnologies": [
                          "Rust"
                        ],
                        "gaps": [
                          {
                            "skill": "Rust",
                            "severity": "DEALBREAKER"
                          }
                        ]
                      }
                    },
                    "decisionCard": {
                      "verdict": "PROCEED",
                      "hrSummary": "...",
                      "technicalSummary": "...",
                      "gapDetail": {
                        "overallVerdict": "PROCEED",
                        "technologyFitScore": 0.85,
                        "gaps": [
                          {
                            "dimension": "Rust",
                            "severity": "SIGNIFICANT"
                          }
                        ]
                      }
                    },
                    "interviewQuestions": {
                      "stage": "INTERVIEW_TECHNICAL",
                      "questions": [
                        {
                          "question": "Explain X",
                          "priority": "MUST_ASK"
                        }
                      ]
                    },
                    "notObservable": [
                      "Communication quality"
                    ],
                    "pipelineStageHistory": [
                      {
                        "stage": "APPLIED",
                        "movedAt": "..."
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "Application not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Get application FULL details",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/hr/{appId}/decision": {
      "patch": {
        "description": "Marks an application as SHORTLISTED, REJECTED, or REVIEWED.",
        "operationId": "ApplicantsController_applyDecision",
        "parameters": [
          {
            "name": "appId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "app_123",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyDecisionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Decision applied successfully"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Apply review decision",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/hr/{appId}/scorecard": {
      "get": {
        "description": "Generates a printable HTML scorecard for external sharing.",
        "operationId": "ApplicantsController_getScorecard",
        "parameters": [
          {
            "name": "appId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "app_123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTML scorecard generated"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Export candidate scorecard (HTML)",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/hr/{appId}/stage": {
      "patch": {
        "description": "Moves candidate forward in hiring pipeline.\n\nValidates transitions and may generate interview questions.",
        "operationId": "ApplicantsController_advanceApplicationStage",
        "parameters": [
          {
            "name": "appId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "app_123",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceStageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stage updated successfully"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Advance pipeline stage",
        "tags": [
          "Applications"
        ]
      }
    },
...
  },
  "info": {
    "title": "a16zero Employer API",
    "description": "Backend MVP for Employer platform features and Account Abstraction auth verification.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "internal-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-key",
        "description": "Internal API key"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "ApplyDecisionDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "SHORTLISTED",
              "REJECTED",
              "REVIEWED"
            ],
            "description": "The decision status for the application",
            "example": "SHORTLISTED"
          }
        },
        "required": [
          "status"
        ]
      },
      "AdvanceStageDto": {
        "type": "object",
        "properties": {
          "stage": {
            "type": "string",
            "enum": [
              "APPLIED",
              "REVIEWED",
              "SHORTLISTED",
              "INTERVIEW_HR",
              "INTERVIEW_TECHNICAL",
              "INTERVIEW_FINAL",
              "OFFER",
              "HIRED",
              "REJECTED"
            ],
            "example": "INTERVIEW_HR",
            "description": "Next pipeline stage"
          },
          "note": {
            "type": "string",
            "example": "Strong communication skills",
            "description": "Optional HR note"
          }
        },
        "required": [
          "stage"
        ]
      },
...
    }
  }
}